+++ packages/transforms/filter-schema/package.json
@@ -1,5 +1,5 @@
 {
-    "name": "@graphql-mesh/transform-filter",
+    "name": "@graphql-mesh/transform-filter-schema",
     "version": "0.1.4",
     "sideEffects": false,
     "main": "dist/index.cjs.js",
