+++ packages/transforms/filter-schema/yaml-config.graphql
@@ -1,3 +1,3 @@
 extend type Transform {
-  filter: [String]
+  filterSchema: [String]
 }
